# src/main/java/io/swagger/api/ServicesApi.java
/**
 * NOTE: This class is auto generated by the swagger code generator program (2.3.1).
 * https://github.com/swagger-api/swagger-codegen
 * Do not edit the class manually.
 */
package io.swagger.api;

import javax.validation.Valid;

import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;

import io.swagger.annotations.Api;
import io.swagger.annotations.ApiOperation;
import io.swagger.annotations.ApiParam;
import io.swagger.annotations.ApiResponse;
import io.swagger.annotations.ApiResponses;
import io.swagger.model.Calculator;
import io.swagger.model.Result;


@Api(value = "services", description = "the services API")
public interface ServicesApi {

    @ApiOperation(value = "calculation", nickname = "calculationPOST", notes = "", response = Result.class, tags={ "calculator-REST", })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "OK", response = Result.class),
        @ApiResponse(code = 201, message = "Created"),
        @ApiResponse(code = 401, message = "Unauthorized"),
        @ApiResponse(code = 403, message = "Forbidden"),
        @ApiResponse(code = 404, message = "Not Found") })
    @RequestMapping(value = "/services/calculadoraAES/v1/",
        produces = { "application/xml", "application/json" }, 
        consumes = { "application/xml", "application/json" },
        method = RequestMethod.POST)
    
    ResponseEntity<Result> calculationPOST(@ApiParam(value = "calculator" ,required=true )  @Valid @RequestBody Calculator calculator);

}