# swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Operaciones Basicas Calculadora (+, -, *, / )",
    "version": "1.0.0",
    "title": "Swagger Petstore",
    "termsOfService": "http://swagger.io/terms/",
    "contact": {
      "name": "Gisella Garcia",
      "email": "garciar-g@javeriana.edu.co"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host": "localhost:8080",
  "basePath": "/",
  "tags": [
    {
      "name": "calculator-REST",
      "description": "Calculator Operations"
    }
  ],
  "consumes": [
    "application/xml",
    "application/json"
  ],
  "produces": [
    "application/xml",
    "application/json"
  ],
  "paths": {
    "/services/calculator/v1/operations": {
      "post": {
        "tags": [
          "calculator-REST"
        ],
        "summary": "calculation",
        "operationId": "calculationPOST",
        "consumes": [
          "application/xml",
          "application/json"
        ],
        "produces": [
          "application/xml",
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "calculator",
            "description": "calculator",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Calculator"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Result"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    }
  },
  "definitions": {
    "Calculator": {
      "type": "object",
      "properties": {
        "typeOperation": {
          "type": "string",
          "description": "Tipo de operacion a realizar",
          "enum": [
            "+",
            "-",
            "*",
            "/"
          ]
        },
        "value1": {
          "type": "number",
          "format": "double"
        },
        "value2": {
          "type": "number",
          "format": "double"
        }
      }
    },
    "Result": {
      "type": "object",
      "properties": {
        "mensaje": {
          "type": "string"
        },
        "valueResult": {
          "type": "number",
          "format": "double"
        }
      }
    }
  }
}